refactor: move extension functions to DexKitHelper class
