+++ app/src/main/java/com/gswxxn/camerasnap/dexkit/camera/CameraSnapMembersFinder.kt
@@ -3,11 +3,11 @@ package com.gswxxn.camerasnap.dexkit.camera
 import com.gswxxn.camerasnap.dexkit.CameraMembers
 import com.gswxxn.camerasnap.dexkit.base.BaseFinder
 import com.gswxxn.camerasnap.hook.CameraHooker
-import com.gswxxn.camerasnap.hook.CameraHooker.getMethodInstance
-import com.gswxxn.camerasnap.hook.CameraHooker.uniqueFindMethodInvoking
-import com.gswxxn.camerasnap.hook.CameraHooker.uniqueFindMethodUsingField
 import com.gswxxn.camerasnap.utils.DexKitHelper
 import com.gswxxn.camerasnap.utils.DexKitHelper.findFieldUsingByMethod
+import com.gswxxn.camerasnap.utils.DexKitHelper.uniqueFindMethodInvoking
+import com.gswxxn.camerasnap.utils.DexKitHelper.uniqueFindMethodUsingField
+import com.gswxxn.camerasnap.utils.DexKitHelper.getMethodInstance
 import com.highcapable.yukihookapi.hook.factory.field
 import com.highcapable.yukihookapi.hook.factory.method
 import com.highcapable.yukihookapi.hook.factory.toClass
