+++ app/src/main/java/com/gswxxn/camerasnap/dexkit/camera/SettingsMembersFinder.kt
@@ -3,9 +3,9 @@ package com.gswxxn.camerasnap.dexkit.camera
 import com.gswxxn.camerasnap.dexkit.CameraMembers
 import com.gswxxn.camerasnap.dexkit.base.BaseFinder
 import com.gswxxn.camerasnap.hook.CameraHooker
-import com.gswxxn.camerasnap.hook.CameraHooker.uniqueFindMethodInvoking
-import com.gswxxn.camerasnap.hook.CameraHooker.getMethodInstance
+import com.gswxxn.camerasnap.utils.DexKitHelper.getMethodInstance
 import com.gswxxn.camerasnap.utils.DexKitHelper
+import com.gswxxn.camerasnap.utils.DexKitHelper.uniqueFindMethodInvoking
 import com.highcapable.yukihookapi.hook.factory.method
 import com.highcapable.yukihookapi.hook.factory.toClass
 import com.highcapable.yukihookapi.hook.type.java.IntType
